+++ src/main.ts
@@ -1,4 +1,5 @@
 import './data/beta032Content'
+import './data/rulesCurrent'
 import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'